DONE FOR THE DAY

+++ client/dispatcher.py
@@ -25,14 +25,17 @@ class RequestDispatcher:
         :param queue: The queue to get requests from.
         """
         self.queue = queue
-        channel = grpc.insecure_channel('localhost:50051')
+        options = [
+            ('grpc.max_send_message_length', 1024 * 1024 * 1024),
+            ('grpc.max_receive_message_length', 1024 * 1024 * 1024)]
+        channel = grpc.insecure_channel('localhost:50051', options=options)
         self.stub = file_sync_pb2_grpc.FileSyncStub(channel)
         self.local_db = ClientDatabase()
         self.handlers = {
             EventType.CREATED: self.file_created,
             EventType.DELETED: self.file_deleted,
-            # EventType.MODIFIED: self.file_modified,
-            # EventType.MOVED: self.file_moved,
+            EventType.MODIFIED: self.file_modified,
+            EventType.MOVED: self.file_moved,
             # EventType.ROUTINE_CHECK: self.routine_check
         }
 
@@ -58,6 +61,8 @@ class RequestDispatcher:
                 break
             except PermissionError:
                 time.sleep(1)
+            except FileNotFoundError:
+                return
 
         file_hash = hashlib.sha256(file_data).hexdigest()
         timestamp = Timestamp()
@@ -73,8 +78,41 @@ class RequestDispatcher:
         Handle the file deleted event.
         """
         file_info = self.local_db.get_file(event.src_path)
+        if not file_info:
+            return
         result = self.stub.delete_file(file_sync_pb2.FileRequest(user_id="1", file_id=file_info["file_id"])).value
         if not result:
             logging.warning(f"Failed to delete file with id: {file_info['file_id']}")
             return
-        self.local_db.delete_record(event.src_path)
+
+    def file_modified(self, event: Event):
+        """
+        Handle the file modified event.
+        """
+        file_info = self.local_db.get_file(event.src_path)
+        file_changes = []
+        try:
+            with open(event.src_path, "rb") as file:
+                for file_part in self.stub.get_file_hashes(
+                        file_sync_pb2.FileRequest(user_id="1", file_id=file_info["file_id"])):
+                    file.seek(file_part.offset)
+                    file_part_content = file.read(file_part.size)
+                    if file_part.hash != hashlib.sha256(file_part_content).hexdigest():
+                        file_changes.append(file_sync_pb2.FilePart(offset=file_part.offset, size=file_part.size,
+                                                                   data=file_part_content))
+                self.stub.sync_file(file_sync_pb2.FileSyncRequest(user_id="1", file_id=file_info["file_id"],
+                                                                  parts=file_changes))
+                file.seek(0)
+                new_hash = hashlib.sha256(file.read()).hexdigest()
+                self.local_db.update_file_hash(file_info["file_id"], new_hash)
+                logging.info(f"Updated file with id: {file_info['file_id']}, new hash: {new_hash}")
+        except FileNotFoundError:
+            pass
+
+    def file_moved(self, event: Event):
+        """
+        Handle the file moved event.
+        """
+        file_id = self.local_db.get_file(event.src_path)["file_id"]
+        self.local_db.update_file_name(event.src_path, event.dest_path)
+        self.stub.update_file_name(file_sync_pb2.UpdateFileName(user_id="1", file_id=file_id, new_name=event.dest_path))
